+++ src/main.py
@@ -7,13 +7,25 @@ from data import Data
 
 
 class DataHandler:
+    """
+    Diese Klasse enthält Methoden zum Speichern, Laden und Bearbeiten von Daten in einer JSON-Datei.
+    """
     @staticmethod
     def save_json():
+        """
+        Speichert die Daten in der "price-control.json" Datei.
+        """
         with open("price-control.json", "w") as file:
             json.dump(Data.data, file)
 
     @classmethod
     def insert_data(cls, site: str, name: str, price: float):
+        """
+        Fügt zur richtigen Webseite und Namen das Datum mit dem aktuellen Preis.
+        :param site: Webseite, zu der die Daten gehören.
+        :param name: Der Name des Produkts.
+        :param price: Der Preis des Produkts
+        """
         today = datetime.now()
         formatted_date = today.strftime("%d.%m.%Y")
 
@@ -28,6 +40,10 @@ class DataHandler:
 
     @classmethod
     def load_json(cls, autostart: bool):
+        """
+        Lädt die JSON Datei "price-control.json".
+        :param autostart: Ein Boolean der angibt, ob es sich um einen Autostart handelt.
+        """
         file_name = "price-control.json"
 
         if os.path.isfile(file_name):
@@ -35,8 +51,8 @@ class DataHandler:
                 Data.data = json.load(file)
 
             if not autostart:
-                noti = notification.Notification
-                noti.create_window(noti(), False, [])
+                noti = notification.Notification()
+                noti.run(False, [])
 
             cls.save_json()
         else: